Allow values other than vec to be used in repeaters
Since we were always borrowing the value before iterating over it,
repetitions essentially required `T where for<'a> &'a T: IntoIterator`,
which is pretty much only true for `Vec`. By default we now own values
passed into a repetition. `&ident` is valid as a way to retain the old
behavior, but it's also usually easy to just get a reference elsewhere
since the value must be assigned to a local variable.