Use local_inner_macros to resolve all helper macros within $crate
This fixes the following error when using Rust 2018 style macro imports.

    use quote::quote;

    error: cannot find macro `quote_spanned!` in this scope
     --> src/main.rs:4:5
      |
    4 | /     quote! {
    5 | |     }
      | |_____^

The `local_inner_macros` modifier resolves all macro invocations made
from within that macro as helpers in the same crate. So if `quote!`
expands to an invocation of `quote_spanned!` then this would be resolved
as `$crate::quote_spanned!` rather than requiring the caller to have
`quote_spanned` in scope.

The attribute is ignored by pre-2018 compilers so quote will continue to
work as normal with #[macro_use].

In the future when dropping compatibility with pre-2018 compilers we can
remove the `local_inner_macros` modifier and use our own explicit
`$crate::` prefixes on invocations of helper macros.